[REF] CRM_Core - Move shared smarty functionality into a trait
Simply moves duplicate functions, no functional change.
